requirements.txt
fastapi
uvicorn
pymongo
google-genai
python-dotenv
aiofiles
